+++ src/storage/AnimCore.ts
@@ -370,7 +370,7 @@ export let AnimCore = class AnimCore {
 		});
 
 		for (const anim of Object.values(validAnimations)) {
-			if (!anim.length) return;
+			if (!anim.length) continue;
 
 			const sequence = new Sequence({ inModuleName: 'pf2e-graphics', softFail: !dev });
 
@@ -381,7 +381,7 @@ export let AnimCore = class AnimCore {
 				);
 			}
 
-			await sequence.play({ preload: true, local: true });
+			sequence.play({ preload: true, local: true });
 		}
 	}
 
@@ -408,7 +408,6 @@ export let AnimCore = class AnimCore {
 			'damage-roll',
 			'place-template',
 			'action',
-			'spell-cast',
 			'toggle',
 			'effect',
 			'self-effect',