+++ CHANGELOG.md
@@ -10,9 +10,21 @@ and this project adheres to [Semantic Versioning](<https://semver.org/spec/v2.0.
 
 ## [Unreleased]
 
+## [0.8.3] - 2024-08-28
+
 ### Added
 
-- Triggers being shown in Animation History
+- Triggers being shown in Animation History.
+- Rudimentary support for rulesSelections as predicates (ex. `effect:rulesSelections:aidBonus:-1`).
+
+### Changed
+
+- **Removed `spell-cast` trigger in favor of `action`.**
+
+### Fixed
+
+- Stabilize and Aid animations.
+- Multiple animations exiting early instead of playing every triggered animation.
 
 ## [0.8.2] - 2024-08-28
 
@@ -388,7 +400,8 @@ and this project adheres to [Semantic Versioning](<https://semver.org/spec/v2.0.
 - Moved internal structure
 - Pre-release versions to be semver compliant
 
-[Unreleased]: https://github.com/MrVauxs/pf2e-graphics/compare/v0.8.2...HEAD
+[Unreleased]: https://github.com/MrVauxs/pf2e-graphics/compare/v0.8.3...HEAD
+[0.8.3]: https://github.com/MrVauxs/pf2e-graphics/compare/v0.8.2...v0.8.3
 [0.8.2]: https://github.com/MrVauxs/pf2e-graphics/compare/v0.8.1...v0.8.2
 [0.8.1]: https://github.com/MrVauxs/pf2e-graphics/compare/v0.8.0...v0.8.1
 [0.8.0]: https://github.com/MrVauxs/pf2e-graphics/compare/v0.7.4...v0.8.0